bug fix
Fixed error when rnafold predicts no paired bases, negative integer results in calc_overhangs function

+++ src/cpp_functions.cpp
@@ -70,7 +70,7 @@ DataFrame mergePileups(std::vector<int> start, std::vector<int> end, std::vector
 // [[Rcpp::export]]
 DataFrame get_nearby(IntegerVector f_start, IntegerVector f_end, IntegerVector r_start, IntegerVector r_end, int max_dist, int dfsize){
   // To be deprecated - replaced with 2 lines of dplyr functions
-  
+
   // Calculate the distance between overlapping read pairs
 
   // Calculate the distance between overlapping read pairs
@@ -368,7 +368,7 @@ DataFrame getPileupsMap(std::vector<int> dtpos, std::vector<int> dtcount,
       int r2_start = start_r2[i];
       int r2_end = end_r2[i];
       int r2_dupes = dupe_count_r2[i];
-      
+
 
 
       //set read length to be read end - read start
@@ -508,14 +508,14 @@ DataFrame calc_overhangs(std::vector<int> r1_start, std::vector<int> r1_end,
   int vector_length = r1_start.size();
   int MEMORY_SIZE = 9;
 
-  std::vector<int> proper_count;
-  std::vector<int> improper_count;
+  std::vector<double> proper_count;
+  std::vector<double> improper_count;
   proper_count.reserve(MEMORY_SIZE);
   improper_count.reserve(MEMORY_SIZE);
 
   for(int i = -4; i <= 4; i++){
-    int pcount = 0;
-    int icount = 0;
+    double pcount = 0;
+    double icount = 0;
 
     for (int j = 0; j < vector_length; j++) {
       int new_r2_start = r2_start[j] + i; // Shift r2_start
@@ -843,7 +843,7 @@ DataFrame get_phased_dist(IntegerVector startX, IntegerVector endX,
 // [[Rcpp::export]]
 DataFrame make_count_table(std::vector<int> fdt_start, std::vector<int> fdt_end, std::vector<int> fwidth,
                            std::vector<int> rdt_start, std::vector<int> rdt_end, std::vector<int> rwidth) {
-  
+
   int res_size = 27;
 
   std::vector<int> overlap_res(res_size);
